Add PlaceMark method to gameboard, define players on gameflow
PlaceMark method will get the mark from the player and will
put it on the gameboard matrix.

Define both players in GameFlow where we are gonna control
which is playing.

PlayerFactory saves as a variable the mark that the player
will put in the gameboard.
number 0 equals to mark 'O'.
number 1 equals to mark 'X'.