Add check 3 marks in a row.
_ArrayIndexToMatrix in GameBoard.
 - Converted repetitive code into a function
   to get matrix index based on array index.

Check3InARow in GameBoard Object.
 - checks if we have 3 in columns, rows or diagonal
   and return true if there is 3 marks in a row.
 - Added to the event click in GameFlow Object.
 - To do so this functions calls the next functions:

_CheckRow in GameBoard Object.
 - check if there are 3 times the same mark in the rows.

_CheckColumn in GameBoard Object.
 - check if there are 3 times the same mark in the columns.

_CheckDiagonal in GameBoard Object.
 - check if there are 3 times the same mark in the diagonal.

Extra change

Change how the matrix is used, it was used like this
matrix[x][y] and now will be used like this matrix[y][x]
so the x is equivalent to moving right and left and
y is equivant to move up and down.
previously we had
  x = Math.floor(index/3);
  y = index%3;
Now we have
  y = Math.floor(index/3);
  x = index%3;
so both times we moved left to right first and then up to bottom.
It's more of a conceptual change.